requirements.txt
matplotlib==3.3.4
numpy==1.24.4
pillow==10.4.0
scikit-learn==1.3.2
tqdm==4.67.1
torch==1.8.1+cu101
torchvision==0.9.1+cu101